requirements.txt
black
dllist  # for task 2
mypy
pre-commit
pytest
